이미지 업로드 URL 취득 함수의 공통화
