+++ app/me/setting/action.ts
@@ -60,13 +60,3 @@ export async function updateProfile(
     redirect("/me");
   }
 }
-
-export async function getUploadUrl() {
-  const response = await fetch(constants.CF_DIRECT_UPLOAD_URL, {
-    method: "POST",
-    headers: { Authorization: `Bearer ${process.env.CF_API_TOKEN}` },
-  });
-  const data = await response.json();
-
-  return data;
-}
